separated out dialog helpers
